move TestRunner into lib/src
